Speed up player movement
If we compare the _Before_ and _After_ videos in the PR we see a marked difference in the speed of player movement in the _After_ video.

This is due to a switch from `setInterval` to `requestAnimationFrame`.

* [requestAnimationFrame](https://developer.mozilla.org/en-US/docs/Web/API/Window/requestAnimationFrame)

When doing tasks such as drawing (animation), using `requestAnimationFrame` is (almost always) the preferred (and better) API:

* [Why?](https://stackoverflow.com/questions/38709923/why-is-requestanimationframe-better-than-setinterval-or-settimeout/38709924#38709924)