Replace JS animation with animated image
The handrolled JS animation worked but it was slowing the responsiveness of the game to a crawl.

There's nothing dynamic about the animation so I have created a single aniated PNG from the animation frames using [this site](https://ezgif.com/apng-maker).

The responsiveness of the game is now much improved because the browser just renders the PNG which is self-animating and fetches the PNG in a single network call.